move out of sample folder and rename to components
