+++ tabs/src/Privacy.jsx
