+++ tabs/src/TabConfig.jsx
