+++ tabs/src/components/app/MainContent.jsx
