+++ tabs/src/components/app/views/ConnectedResources.jsx
