+++ tabs/src/components/app/views/RSCList.jsx
