+++ tabs/src/components/app/views/SampleQueries.jsx
