+++ tabs/src/components/style/ConnectedResources.css
