+++ tabs/src/components/style/DocumentationLinks.css
