+++ tabs/src/components/style/MainContent.css
