+++ tabs/src/components/style/SampleQueries.css
