+++ tabs/src/components/style/TabConfig.css
