+++ tabs/src/index.css
@@ -1,11 +1,11 @@
-@import url("components/sample/style/TabConfig.css");
-@import url("components/sample/style/ConnectedResources.css");
-@import url("components/sample/style/MainContent.css");
-@import url("components/sample/style/TableOfContents.css");
-@import url("components/sample/style/QueryRunner.css");
-@import url("components/sample/style/Tab.css");
-@import url("components/sample/style/DocumentationLinks.css");
-@import url("components/sample/style/SampleQueries.css");
+@import url("components/style/TabConfig.css");
+@import url("components/style/ConnectedResources.css");
+@import url("components/style/MainContent.css");
+@import url("components/style/TableOfContents.css");
+@import url("components/style/QueryRunner.css");
+@import url("components/style/Tab.css");
+@import url("components/style/DocumentationLinks.css");
+@import url("components/style/SampleQueries.css");
 * {
   box-sizing: border-box;
   margin: 0px;
